Add new Ruby 2.6 image with Debian Bullseye
This commit adds a new Ruby 2.6 image running Debian Bullseye.

On the MarkOps team we use a library called Nokogiri which now supports
Apple Silicon with the caveat that you have to have a version of `glibc
>= 2.9`. This version only exists in Debian Bullseye so we're upgrading
in order to better support Apple Silicon on the MarkOps team.

I opted to create additional images instead of modifying the existing
one so that other teams are not affected by this change.