Change how postgresql-client-9.6 and qt5 are installed in Bullseye (#49)
This commit changes how qt5-default and postgresql-client-9.6 are
installed on our Ruby Bullseye image.

When running this image on Apple Silicon we ran into two issues: We needed
the postgres-sql-client from the official Postgres repository and we
needed to install the qt5-default packages individually. This image now
builds successfully on Apple Silicon (tested locally by devs with Apple
Silicon).

Unfortunately because neither `curl` or `wget` is initially installed, I
had to run the update command twice once for the Debian packages and
once for the Postgres packages. This also means we're running an
install twice respectively. We'll remove this code once we get postgres
updated to a version in the Debian repo (>= 13).